+++ .env.example
@@ -1,0 +1,3 @@
+BSKY_USERNAME=""
+BSKY_PASSWORD=""
+RUST_LOG=info
