+++ Cargo.toml
@@ -9,7 +9,11 @@ edition = "2021"
 hyper = { version = "0.14", features = ["full"] }
 hyper-tls = "0.5"
 tokio = { version = "1", features = ["full"] }
+
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
 csv = "1.1"
+
 dotenv = "0.15.0"
+log = "0.4"
+env_logger = "0.10"
